split up build.js into 2 parts, 1 for scrolling, 1 for paywall building
